backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
celery
redis
litellm
requests
pydantic
pydantic-settings
python-dotenv
PyGithub
semgrep
